Migrate build back to setuptools (#139)
The build process had changed to poetry a few versions back to try out
something new and to simplify the build process. After careful
consideration this commit migrates back to setuptools but still takes
advantage of PEP 517 as the build process.

This also migrates the project layout to a src layout to better ensure
that testing is done against a built produce rather than on the source
code which might contain setups not present in a wheel (package data not
present etc). While not a necessary change for migration back to
setuptools this aligns with a few of my other libraries.