Add black, isort, and mypy (#140)
Changes the sole linting tool from pycodestyle to a combination of
black, isort, and mypy. The isort tool makes it simpler to manage
imports in a common manner and black's auto formatting is very helpful
in ensuring a consistent codebase.

Mypy is enabled in a limited fashion to start introducting type
annotations to the code base. The py.typed is also included in the sdist
and wheel for users of this library to pick up for type validation.

Finally these are also added to a pre-commit hook that helps developers
of this library more easily run the sanity checks if they wish to do so.