Remove 1.0.0 beta changes
Remove the changes for the v1.0.0 release as that is now in another
branch for future work. For now this repo will focus on the 0.x release
in order to continue to create bugfixes that are needed.

The v1.x features will still be available as a beta but will continue to
live in a separate branch until it is ready to come back into the main
branch.