Add retry for Receive operations (#208)
* Add retry for Receive operations

Adds a mechanism to retry a WSMan `Receive` operation up to 5 times.
This is useful if the command being run bounces the network connection
leading to the socket being closed. WSMan is able to recover from this
situation as it caches the results for each message id received.

The retry mechanism will timeout based on the same
`reconnection_backoff` option passed into the WSMan client.

* Also handle ConnectTimeout in case the host isn't fully ready yet

* Use .copy() to create a copy of the headers