Add support for multiple PowerShell invocations
Add support for invoking a PowerShell pipeline multiple times. It is
important that the pipeline is closed between invocations to clear up
any server resources and this is kept as a manual step to keep the
normal use cases more efficient in the number of messages sent.

Also added the ability to use `PowerShell` with the `with` statement to
provide a nicer interface for ensuring the pipeline is closed/cleaned up
if this is desired.